Folder_Firecrawl/requirements.txt
firecrawl-py
streamlit
python-dotenv
pandas
pydantic